ref(onboarding): Extract SCM connect sub-components to components directory
Move ProviderPills, ConnectedView, and RepoSelector into
static/app/views/onboarding/components/ to match the existing pattern
where step components live at the root and sub-components live in
components/.

Refs VDY-19
